Renamed `queries.ts` to `hooks.ts`.
