+++ packages/esm-patient-list-app/src/ui-components/create-edit-patient-list/create-edit-list.component.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from 'react-i18next';
 import styles from './create-edit-patient-list.scss';
 import { useLayoutType, showToast, useSessionUser } from '@openmrs/esm-framework';
 import { createPatientList, editPatientList } from '../../api/api-remote';
-import { useCohortTypes } from '../../api/queries';
+import { useCohortTypes } from '../../api/hooks';
 import { OpenmrsCohort, NewCohortData } from '../../api/types';
 
 interface CreateEditPatientListProps {
